Update patch set 2
Patch Set 2:

The devices in the tmo build have a pn547 chip, but it behaves as if it was a pn544. Samsung even shipped a legacy pn544 nfc hal. No amount of fighting getting a firmware to load on init is successful, so NFC was disabled on this build to prevent wakelocks due to the service constantly reloading.

The devices that are part of the hlte build have a bcm2079x which initialized properly.

This is the whole reason these builds are separate.

Patch-set: 2
Reviewer: Kevin F. Haggerty <15173@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0